requirements.txt
xmltodict
PyYAML
Keras==2.0.8
Theano==0.8.2
SQLAlchemy==1.1.9
blessings
github3.py==0.9.6
h5py==2.6.0
javac-parser==0.2.0
jenkspy==0.1.3
more-itertools==2.3
numpy>=1.11.0
python-dateutil==2.6.0
python-Levenshtein==0.12.0
redis==2.10.5
requests==2.13.0
tqdm
